Refactor expression components into separate files for better modularity
Split expression components from single components.tsx file into individual
component files. Each expression type now has its own dedicated file following
React best practices and improving code organization.
