+++ src/views/expressions/UpdateExpression.tsx
@@ -1,0 +1,35 @@
+import React, { ReactNode } from 'react';
+import { ExpressionProps } from './types';
+import { Context } from '../constructs';
+import { dispatchExpression } from './dispatchers';
+
+function renderUnaryExpression(className: string, context: Context): ReactNode {
+  const op = context.node?.get('operator') as ReactNode;
+  const argument = context.child('argument').render(dispatchExpression);
+  className = 'expression ' + className;
+  if (op === 'typeof') {
+    return (
+      <span className={className}>
+        <span className="keyword operator">{op}</span> {argument}
+      </span>
+    );
+  } else if (context.node?.get('prefix')) {
+    return (
+      <span className={className}>
+        <span className="operator">{op}</span>
+        {argument}
+      </span>
+    );
+  } else {
+    return (
+      <span className={className}>
+        {argument}
+        <span className="operator">{op}</span>
+      </span>
+    );
+  }
+}
+
+export const UpdateExpression: React.FC<ExpressionProps> = (props) => {
+  return renderUnaryExpression('update-expression', new Context(props));
+};
